Extract all Exception handling code into its own class
